database/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cmpe492-database"
version = "0.1.0"
description = "PostgreSQL database client for crypto exchange analysis"
requires-python = ">=3.10"
dependencies = [
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=0.19.0",
]

[tool.setuptools]
py-modules = ["db_client", "db_init"]
